requirements.txt
torch
torchvision
timm==0.3.2
tensorboard
pandas
python-decouple
matplotlib
wandb